updates from monkey v1

+++ monkeyKit/src/main/java/com/criptext/comunication/MOKMessageHandler.java
@@ -49,9 +49,30 @@ public class MOKMessageHandler extends Handler {
                             } else if (message.getProps().has("monkey_action")) {
                                 type = message.getProps().get("monkey_action").getAsInt();
                                 message.setMonkeyAction(type);
-                                service.processMessageFromHandler(CBTypes.onNotificationReceived, new Object[]{message});
+                                if(message.getMessage_id()!=null && message.getMessage_id().length()>0 && !message.getMessage_id().equals("0")){
+                                    service.setLastTimeSynced(Long.parseLong(message.getDatetime()));
+                                }
+                                try {
+                                    switch (type) {
+                                        case com.criptext.comunication.MessageTypes.MOKGroupCreate:
+                                            service.processMessageFromHandler(CBTypes.onGroupAdded, new Object[]{message.getProps().get("group_id").getAsString(), message.getProps().get("members").getAsString(), message.getProps().get("info")});
+                                            break;
+                                        case com.criptext.comunication.MessageTypes.MOKGroupNewMember:
+                                            service.processMessageFromHandler(CBTypes.onGroupNewMember, new Object[]{message.getRid(), message.getProps().get("new_member").getAsString()});
+                                            break;
+                                        case com.criptext.comunication.MessageTypes.MOKGroupRemoveMember:
+                                            service.processMessageFromHandler(CBTypes.onGroupRemovedMember, new Object[]{message.getRid(), message.getSid()});
+                                            break;
+                                        case com.criptext.comunication.MessageTypes.MOKGroupJoined:
+                                            service.processMessageFromHandler(CBTypes.onGroupsRecover, new Object[]{message.getMsg()});
+                                            break;
+                                    }
+                                }
+                                catch (Exception e){
+                                    e.printStackTrace();
+                                }
                             } else
-                                service.processMessageFromHandler(CBTypes.onNotificationReceived, new Object[]{message});
+                                service.processMessageFromHandler(CBTypes.onNotificationReceived, new Object[]{message.getMessage_id(), message.getSid(), message.getRid(), message.getParams(), message.getDatetime()});
                         }
                         break;
                     case MessageTypes.MOKProtocolMessageBatch:
@@ -66,26 +87,40 @@ public class MOKMessageHandler extends Handler {
                         break;
                     case MessageTypes.MOKProtocolAck:
                         try {
-                            System.out.println("MOK ack 205");
-                            //The acknowledge message has the id of the successfully sent message in
-                            //the Msg field. We'll use that to update our pending messages list.
-                            service.removePendingMessage(message.getMsg());
-                            service.processMessageFromHandler(CBTypes.onAcknowledgeReceived, new Object[]{message});
+                            if(message.getType().compareTo(MessageTypes.MOKOpen)==0){
+                                service.processMessageFromHandler(CBTypes.onConversationOpenResponse
+                                                , new Object[]{message.getSid()
+                                                        ,message.getProps().get("online").getAsString().compareTo("1")==0
+                                                        ,message.getProps().has("last_seen")?message.getProps().get("last_seen").getAsString():null
+                                                        ,message.getProps().has("last_open_me")?message.getProps().get("last_open_me").getAsString():null});
+                            }
+                            else {
+                                //The acknowledge message has the id of the successfully sent message in
+                                //the Msg field. We'll use that to update our pending messages list.
+                                service.removePendingMessage(message.getMsg());
+                                service.processMessageFromHandler(CBTypes.onAcknowledgeReceived
+                                        , new Object[]{message.getSid()
+                                                ,message.getRid(),message.getMessage_id()
+                                                ,message.getOldId(),message.getStatus() == 52
+                                                ,Integer.parseInt(message.getType())});
+                            }
                         } catch (Exception e) {
                             e.printStackTrace();
                         }
                         break;
                     case MessageTypes.MOKProtocolOpen:{
-                        if(KeyStoreCriptext.getString(service, message.getRid()).compareTo("")==0)
-                            service.sendOpenConversation(message.getRid());
-                        else
-                            System.out.println("MONKEY - llego open pero ya tengo las claves");
-                        //MANDAR AL APP QUE PONGA LEIDO TODOS LOS MENSAJES
-                        service.processMessageFromHandler(CBTypes.onContactOpenMyConversation, new Object[]{message.getSid()});
+                        if(message!=null && !message.isGroupConversation()) {
+                            if (KeyStoreCriptext.getString(service, message.getRid()).compareTo("") == 0)
+                                service.sendOpenConversation(message.getRid());
+                            else
+                                System.out.println("MONKEY - llego open pero ya tengo las claves");
+                            //MANDAR AL APP QUE PONGA LEIDO TODOS LOS MENSAJES
+                            service.processMessageFromHandler(CBTypes.onContactOpenMyConversation, new Object[]{message.getSid()});
+                        }
                         break;
                     }
                     case MessageTypes.MOKProtocolDelete:{
-                        service.processMessageFromHandler(CBTypes.onDeleteReceived, new Object[]{message});
+                        service.processMessageFromHandler(CBTypes.onDeleteReceived, new Object[]{message.getMessage_id(), message.getSid(), message.getRid(), message.getDatetime()});
                         break;
                     }
                     case MessageTypes.MessageSocketConnected:{
@@ -100,11 +135,14 @@ public class MOKMessageHandler extends Handler {
                         break;
                     }
                     case MessageTypes.MOKProtocolGet: {
-                        service.processMessageFromHandler(CBTypes.onNotificationReceived, new Object[]{message});
+                        service.processMessageFromHandler(CBTypes.onNotificationReceived, new Object[]{message.getMessage_id(), message.getSid(), message.getRid(), message.getParams(), message.getDatetime()});
                         break;
                     }
                     case MessageTypes.MOKProtocolSync: {
-                        service.processMessageFromHandler(CBTypes.onNotificationReceived, new Object[]{message});
+                        if(message!=null && message.getMonkeyAction() == MessageTypes.MOKGroupJoined)
+                            service.processMessageFromHandler(CBTypes.onGroupsRecover, new Object[]{message.getMsg()});
+                        else
+                            service.processMessageFromHandler(CBTypes.onNotificationReceived, new Object[]{message.getMessage_id(), message.getSid(), message.getRid(), message.getParams(), message.getDatetime()});
                         break;
                     }
                     default:
